testdrive: rename kafka-verify to kafka-verify-data
To avoid confusion with the relatively recent kafka-verify-commit
action.

Also adjust the module naming for these actions to match their names in
testdrive scripts.
