+++ doc/developer/testdrive.md
@@ -729,7 +729,7 @@ Sets the variable named VAR to the ID of the key schema with which data was
 written. The variable is only set if the format of the key was Confluent Avro
 or Protobuf.
 
-#### `kafka-verify format=avro [sink=... | topic=...] [sort-messages=true] [partial-search=usize]`
+#### `kafka-verify-data format=avro [sink=... | topic=...] [sort-messages=true] [partial-search=usize]`
 
 Obtains the data from the specified `sink` or `topic` and compares it to the expected data recorded in the test. The comparison algorithm is sensitive to the order in which data arrives, so `sort-messages=true` can be used along with manually pre-sorting the expected data in the test.
 
@@ -739,7 +739,6 @@ If `partial-search=usize` is specified, up to `partial-search` records will be r
 
 Verifies that the provided offset (the input data) matches the committed offset
 for the specified consumer group, topic, and partition.
-
 
 #### `headers=<list or object>`
