+++ test/kafka-exactly-once/before-restart.td
@@ -109,17 +109,17 @@ $ kafka-ingest format=avro topic=input schema=${schema}
 {"array":[{"data":{"a":5,"b":6},"time":3,"diff":1}]}
 {"com.materialize.cdc.progress":{"lower":[0],"upper":[4],"counts":[{"time":1,"count":4},{"time":2,"count":2},{"time":3,"count":2}]}}
 
-$ kafka-verify headers=materialize-timestamp format=avro sink=materialize.public.output sort-messages=true
+$ kafka-verify-data headers=materialize-timestamp format=avro sink=materialize.public.output sort-messages=true
 1	{"before": null, "after": {"row": {"a": 1, "b": 1}}}
 1	{"before": null, "after": {"row": {"a": 1, "b": 2}}}
 1	{"before": null, "after": {"row": {"a": 2, "b": 1}}}
 1	{"before": null, "after": {"row": {"a": 3, "b": 1}}}
 
-$ kafka-verify headers=materialize-timestamp format=avro sink=materialize.public.output sort-messages=true
+$ kafka-verify-data headers=materialize-timestamp format=avro sink=materialize.public.output sort-messages=true
 2	{"before": null, "after": {"row": {"a": 11, "b": 11}}}
 2	{"before": null, "after": {"row": {"a": 22, "b": 11}}}
 
-$ kafka-verify headers=materialize-timestamp format=avro sink=materialize.public.output sort-messages=true
+$ kafka-verify-data headers=materialize-timestamp format=avro sink=materialize.public.output sort-messages=true
 3	{"before": null, "after": {"row": {"a": 3, "b": 4}}}
 3	{"before": null, "after": {"row": {"a": 5, "b": 6}}}
