+++ test/kafka-multi-broker/03-after-join.td
@@ -38,7 +38,7 @@ $ kafka-ingest format=avro topic=kafka-multi-broker schema=${schema} timestamp=2
 04-02
 04-03
 
-$ kafka-verify format=avro sink=materialize.public.multi_broker_sink sort-messages=true
+$ kafka-verify-data format=avro sink=materialize.public.multi_broker_sink sort-messages=true
 {"before": null, "after": {"row": {"f1": "01-01" }}}
 {"before": null, "after": {"row": {"f1": "01-02" }}}
 {"before": null, "after": {"row": {"f1": "01-03" }}}
