+++ test/testdrive/consolidation.td
@@ -134,14 +134,14 @@ $ kafka-ingest format=avro topic=nums schema=${nums-schema}
 # verify each transaction separately and sort within each transaction to get
 # deterministic results.
 
-$ kafka-verify headers=materialize-timestamp format=avro sink=materialize.public.nums_sink
+$ kafka-verify-data headers=materialize-timestamp format=avro sink=materialize.public.nums_sink
 1	{"before": null, "after": {"row": {"num": 3}}}
 
-$ kafka-verify headers=materialize-timestamp format=avro sink=materialize.public.nums_sink sort-messages=true
+$ kafka-verify-data headers=materialize-timestamp format=avro sink=materialize.public.nums_sink sort-messages=true
 2	{"before": null, "after": {"row": {"num": 4}}}
 2	{"before": {"row": {"num": 3}}, "after": null}
 
-$ kafka-verify headers=materialize-timestamp format=avro sink=materialize.public.nums_sink sort-messages=true
+$ kafka-verify-data headers=materialize-timestamp format=avro sink=materialize.public.nums_sink sort-messages=true
 3	{"before": null, "after": {"row": {"num": 5}}}
 3	{"before": {"row": {"num": 4}}, "after": null}
 
@@ -153,7 +153,7 @@ $ kafka-verify headers=materialize-timestamp format=avro sink=materialize.public
 #   FORMAT AVRO USING CONFLUENT SCHEMA REGISTRY CONNECTION csr_conn
 #   AS OF 3
 #
-# $ kafka-verify headers=materialize-timestamp format=avro sink=materialize.public.nums_sink
+# $ kafka-verify-data headers=materialize-timestamp format=avro sink=materialize.public.nums_sink
 # 6 {"before": null, "after": {"row": {"num": 5}}}
 
 # Validate that `SUBSCRIBE` is similarly consolidated.
