+++ test/testdrive/disabled/kafka-avro-debezium-transaction.td
@@ -254,12 +254,12 @@ a    b
 
 $ set-regex match=\d{13} replacement=<TIMESTAMP>
 
-$ kafka-verify format=avro sink=materialize.public.data_sink sort-messages=true
+$ kafka-verify-data format=avro sink=materialize.public.data_sink sort-messages=true
 {"before": null, "after": {"row": {"a": 1, "b": 1}}}
 {"before": null, "after": {"row": {"a": 2, "b": 3}}}
 {"before": null, "after": {"row": {"a": 4, "b": 5}}}
 
-$ kafka-verify format=avro sink=materialize.public.data_sink sort-messages=true
+$ kafka-verify-data format=avro sink=materialize.public.data_sink sort-messages=true
 {"before": null, "after": {"row": {"a": 8, "b": 9}}}
 
 #
