+++ test/testdrive/kafka-avro-debezium-sinks.td
@@ -27,7 +27,7 @@
   FORMAT AVRO USING CONFLUENT SCHEMA REGISTRY CONNECTION csr_conn
   ENVELOPE DEBEZIUM
 
-$ kafka-verify format=avro sink=materialize.public.data_sink sort-messages=true
+$ kafka-verify-data format=avro sink=materialize.public.data_sink sort-messages=true
 {"before": null, "after": {"row": {"a": 1, "b": 1}}}
 {"before": null, "after": {"row": {"a": 1, "b": 2}}}
 {"before": null, "after": {"row": {"a": 2, "b": 1}}}
@@ -174,65 +174,65 @@ a  b
 # transaction appear together as one "bundle" in the output. But there is no
 # guarantee on the order within a transaction.
 
-$ kafka-verify headers=materialize-timestamp format=avro sink=materialize.public.non_keyed_sink sort-messages=true
+$ kafka-verify-data headers=materialize-timestamp format=avro sink=materialize.public.non_keyed_sink sort-messages=true
 1	{"before": null, "after": {"row": {"a": 1, "b": 1}}}
 1	{"before": null, "after": {"row": {"a": 2, "b": 2}}}
 
-$ kafka-verify headers=materialize-timestamp format=avro sink=materialize.public.non_keyed_sink sort-messages=true
+$ kafka-verify-data headers=materialize-timestamp format=avro sink=materialize.public.non_keyed_sink sort-messages=true
 2	{"before": null, "after": {"row": {"a": 3, "b": 1}}}
 2	{"before": null, "after": {"row": {"a": 4, "b": 2}}}
 
-$ kafka-verify headers=materialize-timestamp format=avro sink=materialize.public.non_keyed_sink sort-messages=true
+$ kafka-verify-data headers=materialize-timestamp format=avro sink=materialize.public.non_keyed_sink sort-messages=true
 3	{"before": null, "after": {"row": {"a": 1, "b": 7}}}
 
 # Again, compare split by transaction. See comment just above.
 
-$ kafka-verify headers=materialize-timestamp format=avro sink=materialize.public.non_keyed_sink_of_keyed_relation sort-messages=true
+$ kafka-verify-data headers=materialize-timestamp format=avro sink=materialize.public.non_keyed_sink_of_keyed_relation sort-messages=true
 1	{"before": null, "after": {"row": {"a": 1, "b": 1}}}
 1	{"before": null, "after": {"row": {"a": 2, "b": 2}}}
 
-$ kafka-verify headers=materialize-timestamp format=avro sink=materialize.public.non_keyed_sink_of_keyed_relation sort-messages=true
+$ kafka-verify-data headers=materialize-timestamp format=avro sink=materialize.public.non_keyed_sink_of_keyed_relation sort-messages=true
 2	{"before": null, "after": {"row": {"a": 3, "b": 1}}}
 2	{"before": null, "after": {"row": {"a": 4, "b": 2}}}
 
-$ kafka-verify headers=materialize-timestamp format=avro sink=materialize.public.non_keyed_sink_of_keyed_relation sort-messages=true
+$ kafka-verify-data headers=materialize-timestamp format=avro sink=materialize.public.non_keyed_sink_of_keyed_relation sort-messages=true
 3	{"before": null, "after": {"row": {"a": 1, "b": 7}}}
 
 # Again, compare split by transaction. See comment just above.
 
-$ kafka-verify headers=materialize-timestamp format=avro sink=materialize.public.keyed_sink sort-messages=true
+$ kafka-verify-data headers=materialize-timestamp format=avro sink=materialize.public.keyed_sink sort-messages=true
 1	{"a": 1} {"before": null, "after": {"row": {"a": 1, "b": 1}}}
 1	{"a": 2} {"before": null, "after": {"row": {"a": 2, "b": 2}}}
 
-$ kafka-verify headers=materialize-timestamp format=avro sink=materialize.public.keyed_sink sort-messages=true
+$ kafka-verify-data headers=materialize-timestamp format=avro sink=materialize.public.keyed_sink sort-messages=true
 2	{"a": 3} {"before": null, "after": {"row": {"a": 3, "b": 1}}}
 2	{"a": 4} {"before": null, "after": {"row": {"a": 4, "b": 2}}}
 
-$ kafka-verify headers=materialize-timestamp format=avro sink=materialize.public.keyed_sink sort-messages=true
+$ kafka-verify-data headers=materialize-timestamp format=avro sink=materialize.public.keyed_sink sort-messages=true
 3	{"a": 1} {"before": null, "after": {"row": {"a": 1, "b": 7}}}
 
 # Again, compare split by transaction. See comment just above.
 
-$ kafka-verify headers=materialize-timestamp format=avro sink=materialize.public.keyed_sink_of_keyed_relation sort-messages=true
+$ kafka-verify-data headers=materialize-timestamp format=avro sink=materialize.public.keyed_sink_of_keyed_relation sort-messages=true
 1	{"b": 1} {"before": null, "after": {"row": {"a": 1, "b": 1}}}
 1	{"b": 2} {"before": null, "after": {"row": {"a": 2, "b": 2}}}
 
-$ kafka-verify headers=materialize-timestamp format=avro sink=materialize.public.keyed_sink_of_keyed_relation sort-messages=true
+$ kafka-verify-data headers=materialize-timestamp format=avro sink=materialize.public.keyed_sink_of_keyed_relation sort-messages=true
 2	{"b": 1} {"before": null, "after": {"row": {"a": 3, "b": 1}}}
 2	{"b": 2} {"before": null, "after": {"row": {"a": 4, "b": 2}}}
 
-$ kafka-verify headers=materialize-timestamp format=avro sink=materialize.public.keyed_sink_of_keyed_relation sort-messages=true
+$ kafka-verify-data headers=materialize-timestamp format=avro sink=materialize.public.keyed_sink_of_keyed_relation sort-messages=true
 3	{"b": 7} {"before": null, "after": {"row": {"a": 1, "b": 7}}}
 
 # Again, compare split by transaction. See comment just above.
 
-$ kafka-verify headers=materialize-timestamp format=avro sink=materialize.public.multi_keyed_sink sort-messages=true
+$ kafka-verify-data headers=materialize-timestamp format=avro sink=materialize.public.multi_keyed_sink sort-messages=true
 1	{"b": 1, "a": 1} {"before": null, "after": {"row": {"a": 1, "b": 1}}}
 1	{"b": 2, "a": 2} {"before": null, "after": {"row": {"a": 2, "b": 2}}}
 
-$ kafka-verify headers=materialize-timestamp format=avro sink=materialize.public.multi_keyed_sink sort-messages=true
+$ kafka-verify-data headers=materialize-timestamp format=avro sink=materialize.public.multi_keyed_sink sort-messages=true
 2	{"b": 1, "a": 3} {"before": null, "after": {"row": {"a": 3, "b": 1}}}
 2	{"b": 2, "a": 4} {"before": null, "after": {"row": {"a": 4, "b": 2}}}
 
-$ kafka-verify headers=materialize-timestamp format=avro sink=materialize.public.multi_keyed_sink sort-messages=true
+$ kafka-verify-data headers=materialize-timestamp format=avro sink=materialize.public.multi_keyed_sink sort-messages=true
 3	{"b": 7, "a": 1} {"before": null, "after": {"row": {"a": 1, "b": 7}}}
