+++ test/testdrive/kafka-avro-sinks.td
@@ -35,7 +35,7 @@
   FORMAT AVRO USING CONFLUENT SCHEMA REGISTRY CONNECTION csr_conn
   ENVELOPE DEBEZIUM
 
-$ kafka-verify format=avro sink=materialize.public.interval_data_sink sort-messages=true
+$ kafka-verify-data format=avro sink=materialize.public.interval_data_sink sort-messages=true
 {"before": null, "after": {"row": {"interval": [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0]}}}
 {"before": null, "after": {"row": {"interval": [0, 198, 80, 127, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0]}}}
 {"before": null, "after": {"row": {"interval": [0, 58, 175, 128, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0]}}}
@@ -49,7 +49,7 @@ $ kafka-verify format=avro sink=materialize.public.interval_data_sink sort-messa
 #  INTO KAFKA CONNECTION kafka_conn (TOPIC 'testdrive-unnamed-cols-sink-${testdrive.seed}')
 #  FORMAT AVRO USING CONFLUENT SCHEMA REGISTRY CONNECTION csr_conn
 #
-#$ kafka-verify format=avro sink=materialize.public.unnamed_cols_sink
+#$ kafka-verify-data format=avro sink=materialize.public.unnamed_cols_sink
 #{"before": null, "after": {"row": {"column1": 1, "b": 2, "column3": 3}}}
 
 # Test that invented field names do not clash with named columns.
@@ -61,7 +61,7 @@ $ kafka-verify format=avro sink=materialize.public.interval_data_sink sort-messa
 #  INTO KAFKA CONNECTION kafka_conn (TOPIC 'testdrive-clashing-cols-sink-${testdrive.seed}')
 #  FORMAT AVRO USING CONFLUENT SCHEMA REGISTRY CONNECTION csr_conn
 #
-#$ kafka-verify format=avro sink=materialize.public.clashing_cols_sink
+#$ kafka-verify-data format=avro sink=materialize.public.clashing_cols_sink
 #{"before": null, "after": {"row": {"column1": 1, "column1_1": 2, "b": 3, "b2": 4, "b3": 5}}}
 
 # Test date/time types.
@@ -79,7 +79,7 @@ $ kafka-verify format=avro sink=materialize.public.interval_data_sink sort-messa
   FORMAT AVRO USING CONFLUENT SCHEMA REGISTRY CONNECTION csr_conn
   ENVELOPE DEBEZIUM
 
-$ kafka-verify format=avro sink=materialize.public.datetime_data_sink sort-messages=true
+$ kafka-verify-data format=avro sink=materialize.public.datetime_data_sink sort-messages=true
 {"before": null, "after": {"row": {"date": -2440551, "ts": -210863606400000000, "ts_tz": -210863606400000000}}}
 {"before": null, "after": {"row": {"date": 10957, "ts": 946721410111000, "ts_tz": 946714210111000}}}
 {"before": null, "after": {"row": {"date": 10988, "ts": 949399810111000, "ts_tz": 949392610111000}}}
@@ -92,7 +92,7 @@ $ kafka-verify format=avro sink=materialize.public.datetime_data_sink sort-messa
   FORMAT AVRO USING CONFLUENT SCHEMA REGISTRY CONNECTION csr_conn
   ENVELOPE DEBEZIUM
 
-$ kafka-verify format=avro sink=materialize.public.time_data_sink sort-messages=true
+$ kafka-verify-data format=avro sink=materialize.public.time_data_sink sort-messages=true
 {"before": null, "after": {"row": {"time": 0}}}
 {"before": null, "after": {"row": {"time": 3723000000}}}
 {"before": null, "after": {"row": {"time": 3724000000}}}
@@ -118,7 +118,7 @@ $ kafka-verify format=avro sink=materialize.public.time_data_sink sort-messages=
   FORMAT AVRO USING CONFLUENT SCHEMA REGISTRY CONNECTION csr_conn
   ENVELOPE DEBEZIUM
 
-$ kafka-verify format=avro sink=materialize.public.map_sink sort-messages=true
+$ kafka-verify-data format=avro sink=materialize.public.map_sink sort-messages=true
 {"before": null, "after": {"row": {"map": {"a": {"int": 1}, "b": {"int": 2}}}}}
 
 > CREATE MATERIALIZED VIEW list_data (list) AS SELECT LIST[1, 2];
@@ -128,7 +128,7 @@ $ kafka-verify format=avro sink=materialize.public.map_sink sort-messages=true
   FORMAT AVRO USING CONFLUENT SCHEMA REGISTRY CONNECTION csr_conn
   ENVELOPE DEBEZIUM
 
-$ kafka-verify format=avro sink=materialize.public.list_sink sort-messages=true
+$ kafka-verify-data format=avro sink=materialize.public.list_sink sort-messages=true
 {"before": null, "after": {"row": {"list": [{"int": 1}, {"int": 2}]}}}
 
 # Test optional namespace for auto-generated value schema
@@ -143,7 +143,7 @@ $ kafka-verify format=avro sink=materialize.public.list_sink sort-messages=true
 $ schema-registry-verify schema-type=avro subject=testdrive-namespace-value-sink-${testdrive.seed}-value
 {"type":"record","name":"ghi","namespace":"abc.def","fields":[{"name":"before","type":["null",{"type":"record","name":"row","fields":[{"name":"namespace","type":"int"}]}]},{"name":"after","type":["null","row"]}]}
 
-$ kafka-verify format=avro sink=materialize.public.namespace_value_sink sort-messages=true
+$ kafka-verify-data format=avro sink=materialize.public.namespace_value_sink sort-messages=true
 {"before": null, "after": {"row": {"namespace": 1}}}
 
 # Test optional namespaces for autogenerated key and value schemas
@@ -161,7 +161,7 @@ $ schema-registry-verify schema-type=avro subject=testdrive-namespace-key-value-
 $ schema-registry-verify schema-type=avro subject=testdrive-namespace-key-value-sink-${testdrive.seed}-value
 {"type":"record","name":"bar","namespace":"some.neat.class","fields":[{"name":"before","type":["null",{"type":"record","name":"row","fields":[{"name":"a","type":"int"},{"name":"b","type":"int"}]}]},{"name":"after","type":["null","row"]}]}
 
-$ kafka-verify format=avro sink=materialize.public.namespace_key_value_sink sort-messages=true
+$ kafka-verify-data format=avro sink=materialize.public.namespace_key_value_sink sort-messages=true
 {"b": 2} {"before": null, "after": {"row": {"a": 1, "b": 2}}}
 
 # Bad Sinks
