+++ test/testdrive/kafka-avro-upsert-sinks.td
@@ -54,14 +54,14 @@ $ kafka-ingest format=avro topic=upsert-avro key-format=avro key-schema=${upsert
   KEY (key1, key2)
   FORMAT AVRO USING CONFLUENT SCHEMA REGISTRY CONNECTION csr_conn ENVELOPE UPSERT
 
-$ kafka-verify format=avro sink=materialize.public.upsert_input_sink sort-messages=true
+$ kafka-verify-data format=avro sink=materialize.public.upsert_input_sink sort-messages=true
 {"key1": "fisch", "key2": 42} {"key1": "fisch", "key2": 42, "f1": "fish", "f2": 1000}
 {"key1": "fish", "key2": 2} {"key1": "fish", "key2": 2, "f1": "fish", "f2": 1000}
 
 $ kafka-ingest format=avro topic=upsert-avro key-format=avro key-schema=${upsert-keyschema} schema=${upsert-schema}
 {"key1": "fisch", "key2": 42} {"f1": "richtig, fisch", "f2": 2000}
 
-$ kafka-verify format=avro sink=materialize.public.upsert_input_sink
+$ kafka-verify-data format=avro sink=materialize.public.upsert_input_sink
 {"key1": "fisch", "key2": 42} {"key1": "fisch", "key2": 42, "f1": "richtig, fisch", "f2": 2000}
 
 # More complicated scenarios: super keys, consistency input/output
@@ -180,30 +180,30 @@ a  b
 # transaction appear together as one "bundle" in the output. But there is no
 # guarantee on the order within a transaction.
 
-$ kafka-verify headers=materialize-timestamp format=avro sink=materialize.public.input_sink sort-messages=true
+$ kafka-verify-data headers=materialize-timestamp format=avro sink=materialize.public.input_sink sort-messages=true
 1	{"a": 1} {"a": 1, "b": 1}
 1	{"a": 2} {"a": 2, "b": 2}
 
-$ kafka-verify headers=materialize-timestamp format=avro sink=materialize.public.input_sink sort-messages=true
+$ kafka-verify-data headers=materialize-timestamp format=avro sink=materialize.public.input_sink sort-messages=true
 2	{"a": 3} {"a": 3, "b": 1}
 2	{"a": 4} {"a": 4, "b": 2}
 
-$ kafka-verify headers=materialize-timestamp format=avro sink=materialize.public.input_sink sort-messages=true
+$ kafka-verify-data headers=materialize-timestamp format=avro sink=materialize.public.input_sink sort-messages=true
 3	{"a": 1} {"a": 1, "b": 7}
 
 # Again, compare split by transaction. See comment just above.
 
-$ kafka-verify headers=materialize-timestamp format=avro sink=materialize.public.input_sink_multiple_keys sort-messages=true
+$ kafka-verify-data headers=materialize-timestamp format=avro sink=materialize.public.input_sink_multiple_keys sort-messages=true
 1	{"a": 1, "b": 1} {"a": 1, "b": 1}
 1	{"a": 2, "b": 2} {"a": 2, "b": 2}
 
-$ kafka-verify headers=materialize-timestamp format=avro sink=materialize.public.input_sink_multiple_keys sort-messages=true
+$ kafka-verify-data headers=materialize-timestamp format=avro sink=materialize.public.input_sink_multiple_keys sort-messages=true
 2	{"a": 3, "b": 1} {"a": 3, "b": 1}
 2	{"a": 4, "b": 2} {"a": 4, "b": 2}
 
 # missing key denotes DELETE
 
-$ kafka-verify headers=materialize-timestamp format=avro sink=materialize.public.input_sink_multiple_keys sort-messages=true
+$ kafka-verify-data headers=materialize-timestamp format=avro sink=materialize.public.input_sink_multiple_keys sort-messages=true
 3 {"a": 1, "b": 1}
 3	{"a": 1, "b": 7} {"a": 1, "b": 7}
 
@@ -225,14 +225,14 @@ $ kafka-ingest format=avro topic=input-with-deletions schema=${schema}
 {"array":[{"data":{"a":1,"b":1},"time":1,"diff":1}]}
 {"com.materialize.cdc.progress":{"lower":[0],"upper":[2],"counts":[{"time":1,"count":1}]}}
 
-$ kafka-verify headers=materialize-timestamp format=avro sink=materialize.public.input_with_deletions_sink sort-messages=true
+$ kafka-verify-data headers=materialize-timestamp format=avro sink=materialize.public.input_with_deletions_sink sort-messages=true
 1	{"a": 1} {"a": 1, "b": 1}
 
 $ kafka-ingest format=avro topic=input-with-deletions schema=${schema}
 {"array":[{"data":{"a":1,"b":2},"time":2,"diff":1}]}
 {"com.materialize.cdc.progress":{"lower":[2],"upper":[3],"counts":[{"time":2,"count":1}]}}
 
-$ kafka-verify headers=materialize-timestamp format=avro sink=materialize.public.input_with_deletions_sink sort-messages=true
+$ kafka-verify-data headers=materialize-timestamp format=avro sink=materialize.public.input_with_deletions_sink sort-messages=true
 2	{"a": 1} {"a": 1, "b": 2}
 
 # deletion of the "shadowed" input should not cause downstream updates
@@ -247,7 +247,7 @@ $ kafka-ingest format=avro topic=input-with-deletions schema=${schema}
 
 # now we should see a NULL update on the key, which means a DELETE
 
-$ kafka-verify format=avro sink=materialize.public.input_with_deletions_sink sort-messages=true
+$ kafka-verify-data format=avro sink=materialize.public.input_with_deletions_sink sort-messages=true
 {"a": 1}
 
 $ kafka-ingest format=avro topic=input-with-deletions schema=${schema}
@@ -255,7 +255,7 @@ $ kafka-ingest format=avro topic=input-with-deletions schema=${schema}
 {"array":[{"data":{"a":1,"b":2},"time":5,"diff":1}]}
 {"com.materialize.cdc.progress":{"lower":[5],"upper":[6],"counts":[{"time":5,"count":2}]}}
 
-$ kafka-verify headers=materialize-timestamp format=avro sink=materialize.public.input_with_deletions_sink sort-messages=true
+$ kafka-verify-data headers=materialize-timestamp format=avro sink=materialize.public.input_with_deletions_sink sort-messages=true
 5	{"a": 1} {"a": 1, "b": 2}
 
 $ kafka-ingest format=avro topic=input-with-deletions schema=${schema}
@@ -264,7 +264,7 @@ $ kafka-ingest format=avro topic=input-with-deletions schema=${schema}
 
 # removing the occluding input should "reveal" the previous input again
 #
-$ kafka-verify headers=materialize-timestamp format=avro sink=materialize.public.input_with_deletions_sink sort-messages=true
+$ kafka-verify-data headers=materialize-timestamp format=avro sink=materialize.public.input_with_deletions_sink sort-messages=true
 6	{"a": 1} {"a": 1, "b": 1}
 
 # NOT ENFORCED Keys
