+++ test/testdrive/testdrive.td
@@ -97,7 +97,7 @@ contains:unknown catalog item 'UID'
   FORMAT AVRO USING CONFLUENT SCHEMA REGISTRY CONNECTION csr_conn
   ENVELOPE DEBEZIUM
 
-$ kafka-verify format=avro sink=materialize.public.sort_messages_sink sort-messages=true
+$ kafka-verify-data format=avro sink=materialize.public.sort_messages_sink sort-messages=true
 {"before": null, "after": {"row": {"a": 1}}}
 {"before": null, "after": {"row": {"a": 2}}}
 {"before": null, "after": {"row": {"a": 3}}}
@@ -184,7 +184,7 @@ $ kafka-ingest format=avro topic=kafka-ingest-no-partition key-format=avro key-s
   FORMAT AVRO USING CONFLUENT SCHEMA REGISTRY CONNECTION csr_conn
   ENVELOPE DEBEZIUM
 
-$ kafka-verify format=avro sink=materialize.public.kafka_verify_regexp_sink sort-messages=true
+$ kafka-verify-data format=avro sink=materialize.public.kafka_verify_regexp_sink sort-messages=true
 {"before": null, "after": {"row": {"a": "UID"}}}
 {"before": null, "after": {"row": {"a": "UID"}}}
